Round chapters for MusicBrainz using a cumulative offset
MusicBrainz doesn't store track durations down to milliseconds.
At least, not for the digital media medium.
For isn't ideal when using the tracks as chapter offsets for an audiobook.
That's because the rounded or truncated milliseconds can accumulate.
This can result in the chapter offsets being incorrect by a significant number of seconds.

To mitigate this, I've implemented smarter rounding.
Rounding now incorporates a cumulative offset from the result of rounding the previous tracks.
This offset is used to determine whether to round up or down.
Whichever rounding operation brings the cumulative offset closer to 0 is chosen.
This keeps each track's offset within a margin of error of half a second.