Prefer operating on files in place in import scripts
This is more import like behavior for the import scripts.
There are two reasons for this, though.
First, copying things to the tmp directory could potentially lose the benefits of CoW filesystems like Btrfs.
Ideally, these scripts could be run on files that have already been imported.
If they create entirely new files in the view of a CoW filesystem when only updating metadata or performing optimizations, that would be really bad.
By defaulting to behavior that attempts to modify the original file in place as much as possible, this potential pitfall is avoided.
I should probably look more into this to better understand the semantics when moving a file over another in a CoW filesystem.

The second critical reason for this change is to ensure that expensive optimizations aren't lost.
Optimizing the original file in place helps in this situation.
For ACSM conversions to ePUB files, the ePUB will by default be generated in the current working directory instead of in a temporary directory.
This ensures that the EPUB file can be re-used after a failure.